Move metrics to `splitandcache` package
