+++ pkg/frontend/querymiddleware/results_cache_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/mimir/pkg/mimirpb"
+	"github.com/grafana/mimir/pkg/streamingpromql/optimize/plan/splitandcache"
 	"github.com/grafana/mimir/pkg/streamingpromql/requestoptions"
 )
 
@@ -145,19 +146,19 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "@ modifier on vector selector, after end, before maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "metric @ 127"), end: 125000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:                      "@ modifier on vector selector, before end, after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "metric @ 151"), end: 200000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:                      "@ modifier on vector selector, after end, after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "metric @ 151"), end: 125000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:     "@ modifier on vector selector with start() before maxCacheTime",
@@ -168,7 +169,7 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "@ modifier on vector selector with end() after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "metric @ end()"), start: 100000, end: 200000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		// @ modifier on matrix selectors.
 		{
@@ -180,19 +181,19 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "@ modifier on matrix selector, after end, before maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "rate(metric[5m] @ 127)"), end: 125000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:                      "@ modifier on matrix selector, before end, after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "rate(metric[5m] @ 151)"), end: 200000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:                      "@ modifier on matrix selector, after end, after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "rate(metric[5m] @ 151)"), end: 125000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:     "@ modifier on matrix selector with start() before maxCacheTime",
@@ -203,7 +204,7 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "@ modifier on matrix selector with end() after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "rate(metric[5m] @ end())"), start: 100000, end: 200000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		// @ modifier on subqueries.
 		{
@@ -215,19 +216,19 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "@ modifier on subqueries, after end, before maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "sum_over_time(rate(metric[1m])[10m:1m] @ 127)"), end: 125000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:                      "@ modifier on subqueries, before end, after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "sum_over_time(rate(metric[1m])[10m:1m] @ 151)"), end: 200000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:                      "@ modifier on subqueries, after end, after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "sum_over_time(rate(metric[1m])[10m:1m] @ 151)"), end: 125000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		{
 			name:     "@ modifier on subqueries with start() before maxCacheTime",
@@ -238,7 +239,7 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "@ modifier on subqueries with end() after maxCacheTime",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "sum_over_time(rate(metric[1m])[10m:1m] @ end())"), start: 100000, end: 200000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		// offset modifier on vector selectors.
 		{
@@ -250,7 +251,7 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "negative offset on vector selector",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "metric offset -1ms"), end: 125000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		// offset modifier on subqueries.
 		{
@@ -262,7 +263,7 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "negative offset on subqueries",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "sum_over_time(rate(metric[1m])[10m:1m] offset -1ms)"), start: 100000, end: 200000, step: 5},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonModifiersNotCachable,
+			expectedNotCachableReason: splitandcache.NotCachableReasonModifiersNotCachable,
 		},
 		// On step aligned and non-aligned requests
 		{
@@ -274,7 +275,7 @@ func TestIsRequestCachable(t *testing.T) {
 			name:                      "request that is NOT step aligned, with cacheStepUnaligned=false",
 			request:                   &PrometheusRangeQueryRequest{queryExpr: parseQuery(t, "query"), start: 100000, end: 200000, step: 3},
 			expected:                  false,
-			expectedNotCachableReason: NotCachableReasonUnalignedTimeRange,
+			expectedNotCachableReason: splitandcache.NotCachableReasonUnalignedTimeRange,
 			cacheStepUnaligned:        false,
 		},
 		{
