+++ pkg/frontend/querymiddleware/split_and_cache.go
@@ -18,7 +18,6 @@ import (
 	"github.com/grafana/dskit/tenant"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/promql/parser"
 	"go.opentelemetry.io/otel/trace"
@@ -27,15 +26,13 @@ import (
 	apierror "github.com/grafana/mimir/pkg/api/error"
 	"github.com/grafana/mimir/pkg/querier/stats"
 	"github.com/grafana/mimir/pkg/streamingpromql"
+	"github.com/grafana/mimir/pkg/streamingpromql/optimize/plan/splitandcache"
 	"github.com/grafana/mimir/pkg/util/limiter"
 	"github.com/grafana/mimir/pkg/util/spanlogger"
 	"github.com/grafana/mimir/pkg/util/validation"
 )
 
 const (
-	NotCachableReasonUnalignedTimeRange                   = "unaligned-time-range"
-	NotCachableReasonTooNew                               = "too-new"
-	NotCachableReasonModifiersNotCachable                 = "has-modifiers"
 	notCachableReasonModifiersNotCachableFailedParse      = "has-modifiers-failed-parse"
 	notCachableReasonModifiersNotCachableFailedPreprocess = "has-modifiers-failed-preprocess"
 )
@@ -46,40 +43,6 @@ var (
 	defaultMinCacheExtent = (5 * time.Minute).Milliseconds()
 )
 
-type SplitAndCacheMetrics struct {
-	*ResultsCacheMetrics
-
-	SplitQueriesCount              prometheus.Counter
-	QueryResultCacheAttemptedCount prometheus.Counter
-	QueryResultCacheSkippedCount   *prometheus.CounterVec
-}
-
-func NewSplitAndCacheMetrics(reg prometheus.Registerer) *SplitAndCacheMetrics {
-	m := &SplitAndCacheMetrics{
-		ResultsCacheMetrics: NewResultsCacheMetrics("query_range", reg),
-		SplitQueriesCount: promauto.With(reg).NewCounter(prometheus.CounterOpts{
-			Name: "cortex_frontend_split_queries_total",
-			Help: "Total number of underlying query requests after the split by interval is applied.",
-		}),
-		QueryResultCacheAttemptedCount: promauto.With(reg).NewCounter(prometheus.CounterOpts{
-			Name: "cortex_frontend_query_result_cache_attempted_total",
-			Help: "Total number of queries that were attempted to be fetched from cache.",
-		}),
-		QueryResultCacheSkippedCount: promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
-			Name: "cortex_frontend_query_result_cache_skipped_total",
-			Help: "Total number of times a query was not cacheable because of a reason. This metric is tracked for each partial query when time-splitting is enabled.",
-		}, []string{"reason"}),
-	}
-
-	// Initialize known label values.
-	for _, reason := range []string{NotCachableReasonUnalignedTimeRange, NotCachableReasonTooNew,
-		NotCachableReasonModifiersNotCachable} {
-		m.QueryResultCacheSkippedCount.WithLabelValues(reason)
-	}
-
-	return m
-}
-
 // splitAndCacheMiddleware is a MetricsQueryMiddleware that can (optionally) split the query by interval
 // and run split queries through the results cache.
 type splitAndCacheMiddleware struct {
@@ -89,7 +52,7 @@ type splitAndCacheMiddleware struct {
 
 	merger  Merger
 	logger  log.Logger
-	metrics *SplitAndCacheMetrics
+	metrics *splitandcache.SplitAndCacheMetrics
 
 	// Split by interval.
 	splitEnabled  bool
@@ -124,7 +87,7 @@ func newSplitAndCacheMiddleware(
 	logger log.Logger,
 	reg prometheus.Registerer,
 	memoryConsumptionTrackerFactory *limiter.InflightMemoryConsumptionTracker) MetricsQueryMiddleware {
-	metrics := NewSplitAndCacheMetrics(reg)
+	metrics := splitandcache.NewSplitAndCacheMetrics(reg)
 
 	return MetricsQueryMiddlewareFunc(func(next MetricsQueryHandler) MetricsQueryHandler {
 		return &splitAndCacheMiddleware{
